refactor: DependsOn annotation
Now uses annotation "value" and supports an array of them. Added "Util.checkPluginsEnabled" to aid in checking enabled plugins.
