+++ core/src/main/java/com/nisovin/magicspells/castmodifiers/conditions/GriefPreventionIsOwnerCondition.java
@@ -8,10 +8,10 @@ import org.jetbrains.annotations.NotNull;
 import me.ryanhamshire.GriefPrevention.Claim;
 import me.ryanhamshire.GriefPrevention.GriefPrevention;
 
+import com.nisovin.magicspells.util.DependsOn;
 import com.nisovin.magicspells.castmodifiers.Condition;
-import com.nisovin.magicspells.castmodifiers.conditions.util.DependsOn;
 
-@DependsOn(plugin = "GriefPrevention")
+@DependsOn("GriefPrevention")
 public class GriefPreventionIsOwnerCondition extends Condition {
 
 	@Override
