+++ core/src/main/java/com/nisovin/magicspells/castmodifiers/conditions/WorldGuardBooleanFlagCondition.java
@@ -11,10 +11,10 @@ import com.sk89q.worldguard.protection.flags.Flag;
 import com.sk89q.worldguard.protection.flags.BooleanFlag;
 import com.sk89q.worldguard.protection.regions.ProtectedRegion;
 
-import com.nisovin.magicspells.castmodifiers.conditions.util.DependsOn;
+import com.nisovin.magicspells.util.DependsOn;
 import com.nisovin.magicspells.castmodifiers.conditions.util.AbstractWorldGuardCondition;
 
-@DependsOn(plugin = "WorldGuard")
+@DependsOn("WorldGuard")
 public class WorldGuardBooleanFlagCondition extends AbstractWorldGuardCondition {
 
 	protected static final Map<String, BooleanFlag> flags = new HashMap<>();
