+++ core/src/main/java/com/nisovin/magicspells/castmodifiers/conditions/WorldGuardStateFlagCondition.java
@@ -11,10 +11,10 @@ import com.sk89q.worldguard.protection.flags.Flag;
 import com.sk89q.worldguard.protection.flags.StateFlag;
 import com.sk89q.worldguard.protection.regions.ProtectedRegion;
 
-import com.nisovin.magicspells.castmodifiers.conditions.util.DependsOn;
+import com.nisovin.magicspells.util.DependsOn;
 import com.nisovin.magicspells.castmodifiers.conditions.util.AbstractWorldGuardCondition;
 
-@DependsOn(plugin = "WorldGuard")
+@DependsOn("WorldGuard")
 public class WorldGuardStateFlagCondition extends AbstractWorldGuardCondition {
 
 	private static final Map<String, StateFlag> flags = new HashMap<>();
